Inline brokered service factories into the services
